+++ src/main/java/fi/dy/masa/malilib/MaLiLibInitHandler.java
@@ -1,5 +1,6 @@
 package fi.dy.masa.malilib;
 
+import fi.dy.masa.malilib.command.ClientCommandHandler;
 import fi.dy.masa.malilib.config.ConfigManager;
 import fi.dy.masa.malilib.event.InputEventHandler;
 import fi.dy.masa.malilib.event.RenderEventHandler;
@@ -9,6 +10,7 @@ import fi.dy.masa.malilib.hotkeys.IKeybind;
 import fi.dy.masa.malilib.hotkeys.KeyAction;
 import fi.dy.masa.malilib.interfaces.IInitializationHandler;
 import fi.dy.masa.malilib.interfaces.IRenderer;
+import fi.dy.masa.malilib.test.TestCommand;
 import fi.dy.masa.malilib.test.TestInputHandler;
 import fi.dy.masa.malilib.test.TestRenderHandler;
 
@@ -30,6 +32,8 @@ public class MaLiLibInitHandler implements IInitializationHandler
             RenderEventHandler.getInstance().registerTooltipLastRenderer(renderer);
             RenderEventHandler.getInstance().registerWorldPreWeatherRenderer(renderer);
             RenderEventHandler.getInstance().registerWorldLastRenderer(renderer);
+
+            ClientCommandHandler.INSTANCE.registerCommand(new TestCommand());
         }
     }