+++ src/main/java/fi/dy/masa/malilib/test/TestRenderHandler.java
@@ -110,7 +110,7 @@ public class TestRenderHandler implements IRenderer
             MutableText title = Text.empty().append(StringUtils.translateAsText(MaLiLibReference.MOD_ID+".gui.tooltip.test.title"));
             list.addFirst(title);
              */
-            list.add(StringUtils.translateAsText(MaLiLibReference.MOD_ID+".gui.tooltip.test.first"));
+            list.add(StringUtils.translateAsText(MaLiLibReference.ID+".gui.tooltip.test.first"));
         }
     }
 
@@ -119,7 +119,7 @@ public class TestRenderHandler implements IRenderer
     {
         if (MaLiLibConfigs.Test.TEST_CONFIG_BOOLEAN.getBooleanValue())
         {
-            list.add(StringUtils.translateAsText(MaLiLibReference.MOD_ID+".gui.tooltip.test.middle"));
+            list.add(StringUtils.translateAsText(MaLiLibReference.ID+".gui.tooltip.test.middle"));
         }
     }
 
@@ -128,7 +128,7 @@ public class TestRenderHandler implements IRenderer
     {
         if (MaLiLibConfigs.Test.TEST_CONFIG_BOOLEAN.getBooleanValue())
         {
-            list.add(StringUtils.translateAsText(MaLiLibReference.MOD_ID+".gui.tooltip.test.last"));
+            list.add(StringUtils.translateAsText(MaLiLibReference.ID+".gui.tooltip.test.last"));
         }
     }