+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ## Changelog
-- remove `ModInitializer`
+- sync `sakura-ryoko/malilib` 1.21.3-0.22.6
+- tag namespace `mafglib` -> `malilib`
+- update NeoForge version to `21.3.76`
+- add `RuntimePackHelper`